Simpler way to check for error count changes from old to new.
Rather than running multiple `phpcs` execs like we do for the new files,
just grab the `--report=json` from the one `phpcs` exec and compare
it to a similar JSON report from the new files.

It means the code between new and old isn't as parallel as before,
but it reduces the comparison complexity.