Docker docs: include all relevant Ngrok regions (#11735)
I set up a Docker instance for Jetpack today, including Ngrok.

During the Ngrok setup process, the docs included an example containing: 

```
region: eu # only needed for subdomains in Europe
```

Ngrok also offers subdomains in Asia/Pacific and Australia. My Ngrok domain is in Australia and `au` needs to be specified in that case.

#### Changes proposed in this Pull Request:

Adds Asia/Pacific and Australia to the example.