NUX / REST API: add `site_segment` to the site options response
Summary:
We need the site segment after signup to properly configure the NUX for our users.

It is currently being imported directly on the `checklist` API response for customizing the checklist but will be needed elsewhere.

The current use-case is needing to detect which segment a user is in so we can redirect them to different places post-signup depending on the segment. This is for the Page Builder MVP, which redirects to the block editor for Segment 1 (Business) only.

Test Plan: TC tests, manual checking of the endpoint

Reviewers: cainm

Reviewed By: cainm

Tags: #touches_jetpack_files

Differential Revision: D26553-code

This commit syncs r190038-wpcom.