Publicize: fix counting words in posts, take 2. (#11810)
Summary:
This is a follow-up to D25267, and D25799 which reverted the first patch because of performance. We're now back to relying on str_word_count to count the number the words, and while that works well for latin languages, it is problematic with languages with special characters, like Japanese or Cyrillic.

This attempts to fix the issue detailed here, where all the content of a post ends up on Facebook, instead of just an excerpt: 3551-gh-jpop-issues

This time around let's try a simpler approach for counting words, by splitting words depending on common separators: space, !, ?, ;, , .
We can expand that regex in the future to support more / different separators.

Test Plan:
- Prepare your Jetpack site and your WordPress.com sandbox for testing:
- Publish a new post with the content defined in the JPOP issue above.
- Make sure only the first part of the post is pushed to Facebook; at the end of the automatically generated excerpt you should see something like "[ 468 more words ]".

Reviewers: kraftbj, biskobe

Tags: #touches_jetpack_files

Differential Revision: D26167-code

This commit syncs r189761-wpcom.