Gutenframe: Show message if user is logged out (#12135)
When a user is not logged into the Jetpack site, the WordPress.com iframed block editor won't load because the iframe redirects to the login URL which cannot be embedded in an iframe (since it causes a sameorigin error screen). 

This change aborts the default workflow with a wp_die command that displays a help message to the user explaining we need them to be logged in. The message is only displayed in a iframed block editor context (current page is post.php or post-new.php and the param `frame-nonce` is present).