Use WordPress.com as the user agent for external requests
On WordPress.com we change the standard WordPress user-agent string to
identify curl requests as coming from WordPress.com. This isn't the case
on Atomic sites and some hosts block requests that use the standard
WordPress user-agent.

This adds the same filter to the user-agent string as on WordPress.com.