Fix updating wpcachehome (#670)
* Set a transient "setting not changed" when settings not changed.

When modifying wp-config.php to add WPCACHEHOME check that transient as
the replace line function will return false since the line may not have
changed..

* Check if setting not changed when updating WP_CACHE too.

* Remove this line as it's completely incorrect.

For more:
https://wordpress.org/support/topic/instruction-to-modify-cachehome-is-not-clear/

* wp_cache_replace_line returns true no change, not false

It used to return false on fatal errors like permission or file missing
problems as well as if setting hadn't changed. This patch makes it
return true if setting hasn't changed.
I had to change other code to cope with that and reversed other changes
in this patch.