Dashboard Build: Use StaticSiteGeneratorPlugin to build Static HTML (#12381)
`tools/builder/react.js` uses `static.jsx` to produce minimal static markup (HTML) versions of the dashboard (really just "Turn on your JavaScript" notices). It `require()`s the webpack-bundled version of `static.jsx` to that end, and attaches `renderToStaticMarkup()`-produced strings to `window`, which it then writes to HTML files.

Getting rid of the complexity added by passing information by attaching to `window` (and thus requiring `jsdom` etc) is the main objective of this PR. This is achieved by using `StaticSiteGeneratorPlugin`, which is meant for this kind of thing. In a subsequent PR, I hope to get rid of `tools/builder/react.js` pretty much entirely. This will make it much easier to base the build system for Jetpack's React Dashboard on `calypso-build` (see #12072).