Add ?checklist query param to checklist CTA route (#12533)
There are plans to change the location of the checklist in Calypso. It
currently shares the my-plan route.

Once Jetpack ships, we'll be unable to modify the route until a future
version and have no way of knowing whether a particular route intends to
show the My Plan page or the Checklist which happens to exist on the My
Plan route at the moment.

By adding a `?checklist` query parameter which is initially unused,
we'll be able to identify the intent to show the Checklist on the My
Plan page and redirect to the checklist route when it is created.