DNA: make sure that local tags match remote tags in package release
When releasing a new version of a package, one of the steps is to fetch all tags from the Jetpack repo. However, if the tags you have in your local copy don't match the remote tags, you get the following error:
```
From github.com:Automattic/jetpack
 ! [rejected]            7.3.1.1    -> 7.3.1.1  (would clobber existing tag)
```

This error interrupts the release process. To avoid this, let's force fetch tags from the remote copy to avoid mismatchs.