Contact Info & Map Widget: Dynamically update customizer admin form
Display API key errors in the Contact Info & Map widget's admin form on
the Customizer preview page. This requres an Ajax call, which results in
a call to the "has_good_map()" method. The "has_good_map()" method
checks the API key by calling the Google Maps API. If the API returns an
error, that error is displayed in the admin form on the Customizer page.