AMP Compatibility: Remove social share CSS
The CSS for the share icons triggers a "illegal_css_at_rule" AMP validation error because of the use of '-moz-document'. Given AMP uses the `amp-social-share` component, this CSS isn't needed at all. This change tells Jetpack not to enqueue the CSS at all when an AMP page is being viewed.