[not verified] Added Carousel checks for existing objects.
In order to avoid fatal errors and interface lockups this change adds some data checks. In case of a gallery being used with images not actually hosted on the site, opening them in a Carousel results in a fatal error and complete lack of control over the UI. This adds a workaround, but still doesn't really make Carousel usable in this context.