Removed fallback / workaround for themes without wp_footer
After discussion on Github issue "Stats: Update method for inserting JavaScript #3763" it was considered that it was a safe assumption to make that a theme would always have a `wp_footer` action.

This update removes code which was used to provide a fallback / workaround in the case of themes which did not contain a `wp_footer` action.

The workaround consisted of setting a global variable `$rendered_stats_footer` which would get set to true if the `wp_footer` call was successful, if not successful then the WP Stats JS code would get set via a `wp_head` action instead.