Contact Info & Map Widget: Remove Geocoding and add map to form (#12474)
* Contact Info & Map Widget: Remove Geocoding and add map to form

The Google Maps Geocoding API is used to verify that the given address
is valid. However, use of this API has recently caused problems for
users. Users need to enable two Google Maps APIs in order to use this
widget, which isn't user-friendly.

Instead of using the Geocoding API results to verify the address, show
the embedded map in the widget's admin form. The user can verify that the
address is valid and correct by looking at the map. Also, the user will
receive Google Maps' error messages if the API key is not valid.

Do not show the embedded map In the Customizer admin form. The embedded
map isn't necessary in this form because the embedded map is shown in
the Customizer preview.

* Rely on esc_url_raw for remote call

https://codex.wordpress.org/Function_Reference/esc_url_raw

* Clarify update process for the map, and coding standard changes

- A few additional comments help understand when has_good_map is needed.
- Use Yoda conditions, we must

* Simplify addition of query params to Google embed URL.

* Only attempt to display the map when we have info to display it

* Enqueue CSS on widgets page as well

This avoids that the map breaks out of the widget area layout when displayed.

* Display output from API request in widget area when error

- This gives more information to users who entered an API key that is not correct for example, instead of just not displaying any map at all.

* Improve code readability and remove an API call.

 * Move the logic for determining whether to update the instance's
   'goodmap' value from the 'update()' function to a new function,
   'update_goodmap()'.

 * Instead of calling 'has_good_map()' in the 'form()' function, just
   check the instance's 'goodmap' value. We've already called
   'has_good_map()' and saved the result in the instance's 'goodmap'
   value, so no need to call "has_good_map()" again.

* Show the API key error in the widget in the customizer preview

When the Google Maps API returns an error, display it in the widget in
the Customizer preview. The CSS properties for the
'contact-map-api-error' class are intended to match the CSS properties
for the admin 'notice-warning' class.

* Contact Info & Map Widget: Dynamically update customizer admin form

Display API key errors in the Contact Info & Map widget's admin form on
the Customizer preview page. This requres an Ajax call, which results in
a call to the "has_good_map()" method. The "has_good_map()" method
checks the API key by calling the Google Maps API. If the API returns an
error, that error is displayed in the admin form on the Customizer page.

* Contact Info: fix linting warnings in js file

* Contact Info CSS: update version so cache can be busted

Addresses https://github.com/Automattic/jetpack/pull/12474#pullrequestreview-267077304

* Do not output notice container when no map has been set up yet

Addresses https://github.com/Automattic/jetpack/pull/12474#pullrequestreview-267077304

* Contact Info & Map Widget: Protect the Ajax request

Use a nonce to verify that the Ajax request is valid. Also check that
the user is allowed to customize before checking the API key from the
Ajax request.

* Fix stylesheet enqueue problem

Enqueue the contact info stylesheet on widgets.php using the
'admin_enqueue_scripts' hook since widgets.php is an admin page.


Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>