Connection Flow: fix calypso_env
This moves the calypso_evn logic from the connection url method and into
the authorization url method. This allows the logic to be shared between
different connection flows.

For example, the new ifram connection flow will now respect the calypso_env parameter
and allow us to test the flow in different calypso environments.