Add support for filtering meta data on post lookup (#13538)
Summary:
> Note: At this point, I'm looking for validation on this approach – if we're happy with it I can fix linting issues, and I plan to add tests.
>

Per https://[private link], we're hoping to be able to filter the metadata in a response down to just what we need.

Because the `meta` work all happens in 	`class.wpcom-json-api.php`, it wasn't possible to modify that data strictly for this endpoint. That file doesn't reference the `fields` parameter at all - that's left to the endpoint to manage.

This leaves us with a couple of implementation questions.

First, it would be fairly trivial to hoist the `filter_response` method from `WPCOM_JSON_API_Post_v1_1_Endpoint` up into `WPCOM_JSON_API_Endpoint`, which would make the meta field filtering available to additional endpoints.

Second, because this is an additional filtering operation, and because we want to keep this API backward-compatible, is it preferable to:

A) When `meta.data.autosave.modified` is specified in the `fields` list, don't only show any other keys under the `meta` namespace? (As a for-instance, `meta.data.autosave.post_ID` would not be present)
B) Keep the existing field specifications as-is, and add a `meta_fields` parameter to the request. A developer could specify something like `&meta_fields=autosave.modified,autosave.post_ID` if they're interested in those fields. If the `meta_fields` query var isn't present, we show all of them.

I personally think option B is preferable, as it's completely backwards-compatible, and combines nicely with the fact that you have to specify a `meta` key in order to fetch `autosave` in the first place. I would appreciate any feedback on this :)

Test Plan:
1) Download this patch to your sandbox
2) Using your favourite API client, request `https://[private link].1/sites/133112643/posts?number=1&fields=title,modified,status,meta&context=edit&meta=autosave&meta_fields=autosave.modified,autosave.post_ID,autosave.ID`. This is a testing site belonging to @malinajirka - alternatively, you could follow his directions to create an autosave post on your own blog:

```
You can create your own “autosave” if you
1. open a published blog post
2. change something
3. wait for about 30s before it’s autosaved
4. leave the blog post without manually saving/publishing the changes.
```

3) Note that the response includes the fields `ID`, `post_ID`, and `modified`.
4) Remove the `meta_fields` param, and note that the response now includes the entire `meta.data.autosafe` object.

Reviewers: malinajirka, migueluy

Reviewed By: migueluy

Subscribers: oguzkocer, malinajirka

Tags: #touches_jetpack_files

Differential Revision: D29715-code

This commit syncs r196943-wpcom.

Co-authored-by: Jeremy Massel <jeremy.massel@automattic.com>