Tracks: Don't track users in dev mode or when opted out (#13698)
 Don't track users in dev mode or when opted out

* Fix all phpcs issues

This will avoid having to bypass the pre-commit hook for all commits to this file

* [not verified] Opt out of tracking in dev mode in legacy branch.

* PHPCS fixes

* Include development mode check within Jetpack::jetpack_tos_agreed.

The TOS check is used for confirming that a site is "Trackable". For sites in development mode or are not active, we should not track them.

* Docblock for is_active_and_not_development function

* Bring cookie check back

Just to be sure. See https://github.com/Automattic/jetpack/pull/13698#discussion_r333960203