Slideshow block: Add image size selector (#13362)
Add an image size selector to the slideshow block that allows users to choose the image size for the gallery inserted in the block.

The selector defaults to large size. Previously, all the images were loaded in full size. This should improve the performance when visiting a page, since the large size will be enough for most users, making necessary to download less MB.