WooCommerce Services: Fix the redirect URL for Woo Services JI… (#13752)
* WooCommerce Services: Use home_url to create redirect url

The redirect URL is created using the value of $_SERVER['REQUEST_URI'], which includes
'wp-admin/'. When admin_url() is used to create the redirect URL, the resulting URL is
incorrect. Use home_url() instead. Also properly unslash and sanitize $_GET['redirect'].

* WooCommerce Services: Use admin_url as the fallback URL

In PR #9239, wp_get_referer() was described as unrealiable because some browsers strip
the 'referer' header. So let's just use admin_url() as the fallback redirect URL if
$_GET['redirect'] is empty.

* WooCommerce Services: Fix PHPCS errors and warnings.

* Fix last PHPCS warning and add file to whitelist.


Co-authored-by: Jeremy Herve <jeremy@jeremy.hu>