Fix logic for choosing  customizer or widgets.php This commit changes the logic for determining whether we are in the customizer or widgets.php. Previously, when accessing widgets from widgets.php, the logic failed to recognize this when the page initialized with no widgets.
This change removes the additional check for .widget-control-actions
when initializing the widgets_shell variable. Instead we
make sure the variable is not undefined, and if it is not undefined, we
make sure it is not empty.

This ensures the correct element is selected, even when there are no
initial active plugins.

See https://github.com/Automattic/jetpack/issues/12302 for details.