Connection: verify_action() needs to return an IXR_Error on failure
The docblock for verify_action() shows that a WP_Error is returned on failure, but this
is incorrect. Any errors generated in verify_action() are passed through the error() method,
which converts them to IXR_Errors. So verify_action() actually returns on IXR_Error on failure.

Make the following changes:
* In verify_action(), check whether verify_secrets() returned an error. If it did, pass that error
  to the error() method to convert it to an IXR_Error.
* Fix the docblock for verify_action().
* Add @return to the verify_secrets() docblock.