Migrates recipe kses to more consistent place
Summary: In D34558 recipe shortcode code was found to be responsible for modifying the global KSES state. This copies the functionality to a core file so this behavior can be removed as it affects JP sites.

Test Plan: Verify KSES output for no security flaws.

Reviewers: kraftbj

Reviewed By: kraftbj

Tags: #touches_jetpack_files

Differential Revision: D34780-code

This commit syncs r199078-wpcom.