Ensure get_mu_plugins is present. (#13991)
get_mu_plugins is within wp-admin/includes/plugin.php, so it is only available via the admin context. The primary purpose of adding this item to the array is to help happiness have that information instantly and is not processed for aggragate purposes. We can leave it out without impacting anything.