Connection: Set the value of user_id in Manager::generate_secrets
Jetpack::generate_secrets sets the value of $user_id when the the $user_id
argument value is false. Move this logic to the Manager::generate_secrets
method. The user id is required for generating the secrets, and this change
ensures that the user id is correctly set when Manager::get_secrets is called.