Connection Owner Deletion Notice: Fix display bug and sanitize… (#14082)
The 'Delete Connection Owner' Notice doesn't display when the connection owner is
about to be deleted. This happens because the arguments to in_array() are an int
and an array of strings. Fix this by converting the elements of the $user_ids_to_delete
array to ints.

Also, santize and unslash the $_REQUEST variable before using it.