Rewind: differenciate provisioning/active states only when needed
Until now, we've shown different messages in the UI when Rewind was being set-up
(i.e until the first backup is complete) vs. after that first backup.
While this can be useful, messaging can be a bit confusing as we would say that
Rewind was being configured, when in fact the site owner already did all the configuring they had to do.

This commit does a few things:
- In some cases, it is actually less confusing to just not differenciate provisioning and active
because we don't necessarily need to highlight that this is a first backup vs. just another backup.
- In other cases, I've updated the wording to match what's currently shown in Calypso
(in the Activity log), to say that a backup was under way instead of saying that something was being configured.
- As soon as Rewind is enabled, and even before the first backup is complete, we now show the Activity log card.
This allows site owners to go to the Activity Log to check on the status of their site, where they will eventually see the first backup listed.

For reference, here is a bit more info about the Provisioning state in Rewind:

> Provisioning state means that the customer has enabled Rewind
> and provided all necessary information but that Rewind isn’t yet protecting their site.
> This may never actually be seen by a site owner
> because it should only exist during the nominally short initialization of the Rewind process.
> A site remains in this state until Rewind has its first complete backup.
> If a site gets stuck preparing the first backup
> this state could stick around for an extended period of time.