Fix logic for choosing between customizer and widgets.php (#13858)
* Fix logic for choosing  customizer or widgets.php
This commit changes the logic for determining whether we are in the
customizer or widgets.php. Previously, when accessing widgets from
widgets.php, the logic failed to recognize this when the page
initialized with no widgets.

This change removes the additional check for .widget-control-actions
when initializing the widgets_shell variable. Instead we
make sure the variable is not undefined, and if it is not undefined, we
make sure it is not empty.

This ensures the correct element is selected, even when there are no
initial active plugins.

See https://github.com/Automattic/jetpack/issues/12302 for details.

* Bumps version in widget admin setup
Bumps version number in widget_admin_setup function to ensure everyone
gets updated file after plugin update.

See https://github.com/Automattic/jetpack/pull/13858#pullrequestreview-319124848