Jetpack Plan: ensure that the class does not block wpcom sync
The class does not exist on WordPress.com, so we should be careful when using it in files that are synchronized with WordPress.com.

The class is currently used in other synced files as well, but in all cases there are early returns before the use of the class when one is on WordPress.com.