Remove broken Pressable nameserver warning https://github.com/Automattic/wpcomsh/pull/363
All Atomic sites are now on Atomic v2, but we are still serving proxying
requests from Pressable's old Peer1 IPs to Automattic. It would be great
to continue warning admins about the need to update their DNS records,
but there is not currently a way to identify those sites short of doing
a DNS lookup for each site on every request.

A cheaper way we could do this is to relay a special HTTP header for
Peer1-proxied requests and only do lookups when we see that. If we don't
want to do that, I think we should simply remove this code now as it is
not worth making all Atomic sites pay for the outdated DNS records of
inattentive admins.

I am removing for now. We can revert and fix later if we want.