Media: Fix the generate_new_filename() method
The file parts of the currently attached media file are obtained in the else
block, but they're needed later regardless of which path through the
if-statement is used. So, obtain the file parts for the current attachment before
the if statement.

Also clean up some PHPCS notices for whitespace and comments.