Media: Fix the notices generated during remote media file updates (#14313)
* Media: Fix the generate_new_filename() method

The file parts of the currently attached media file are obtained in the else
block, but they're needed later regardless of which path through the
if-statement is used. So, obtain the file parts for the current attachment before
the if statement.

Also clean up some PHPCS notices for whitespace and comments.

* Media: Make the get_time_string_from_guid() method static

The get_time_string_from_guid() method is being called statically, which
causes a PHP warning to be generated. Declare the method as static to
silence the warning.

* Media: Check that the input array keys exist with isset()

If the $input array keys 'media', 'media_url', or 'attrs' do not exist,
PHP notices will be generated. Prevent the notices by using isset() to
check that the array keys exist before using them.