Sync: ensure we run the initial sync on new connections
In #14204, we changed the way we initialize the Sync package,
which was a bit too strict and didn't allow for initial syncs
to happen.

Let's make sure we always hook to `jetpack_user_authorized` and
attempt to `do_initial_sync` during connections or reconnections.

There are safeguards in `do_initial_sync` to protect sites that
should sync from doing an initial sync.