Autoloader: Remove the ignored classes
If a package class is loaded before the plugins_loaded action has fired, a PHP
notice is generated by the autoloader. Some classes were known to be loaded before
plugins_loaded was fired. To avoid creating PHP notices for those classes, the
autoloader contains a list of ignored classes. If a class in this list is loaded before
plugins_loaded has fired, a PHP notice is not generated. The intention was to
eventually fix the load order of these classes and remove the list of ignored classes.

We have fixed the load order of the ignored classes, so remove the ignored classes list.
PHP notices will be generated for any package class that is loaded before plugins_loaded
has fired.