Separate block registration from setting the availability
As with other blocks that rely on server side rendering, we need to make
sure that the block is registered so that it is rendered in the frontend
of the site.

This change registers the block during the `init` hook, but keeps the
setting of the block availability to the
`jetpack_register_gutenberg_extensions` hook.