Make the Calendly block registration consistent with OpenTable
In #14464 we're changing the approach to registering the block and
setting its availability, both to fix an issue that meant it wasn't
available in the front end of the site, and to make it consistent with
other blocks.

This change makes the Calendly block follow the same
pattern, by splitting the block registration from the availability
setting, and using the `jetpack_register_gutenberg_extensions` hook.