Config: Add a class that customizes the textdomain in packages
Add a new class, Textdomain_Customizer, that customizes the textdomain of
strings in the packages. It uses the textdomain given in the extra section
of the plugin's composer.json file.

Update Jetpack's composer.json file with the post_install method and
Jetpack's textdomain.

Provide a preliminary unit test for the new class.