[Atomic private sites] Make /sites return is_private based on blog_public site option
Summary: At the moment, `/sites` endpoint will always return false for atomic sites. That's because privacy was never supported on atomic sites. Since this support is coming, we want to base that on blog_public setting. The ultimate fix would be updating SAL, and that's what we're going to do in longer term. A short term fix is to directly consult site_options.

Test Plan:
1. Update `blog_public` to -1 on shadow site of one of your existing atomic sites
1. Go to calypso, inspect call to `/sites` endpoint, confirm it says `is_private: false` and `is_coming_soon: false`
1. Set `wpcom_coming_soon` option on that site to 1
1. Go to calypso, inspect call to `/sites` endpoint, confirm it says `is_private: true` and `is_coming_soon: true`
1. Set `wpcom_coming_soon` option on that site to 0
1. Go to calypso, inspect call to `/sites` endpoint, confirm it says `is_private: true` and `is_coming_soon: false`

Reviewers: jblz

Reviewed By: jblz

Subscribers: jblz

Tags: #touches_jetpack_files

Differential Revision: D38388-code

This commit syncs r202410-wpcom.